+++ frontend/components/AppLayout.tsx
@@ -1,6 +1,6 @@
-import { Container } from "@mui/material";
-import Header from "./Header";
-import Footer from "./Footer";
+import Container from "@mui/material/Container";
+import Header from "@/components/Header";
+import Footer from "@/components/Footer";
 
 interface AppLayoutProps {
   children?: React.ReactNode;
@@ -11,7 +11,7 @@ export default function AppLayout({ children }: AppLayoutProps) {
     <div>
       <Header></Header>
       <Container>{children}</Container>
-      <Footer></Footer>
+      {/* <Footer></Footer> */}
     </div>
   );
 }
